restructure config (#60)
