--- pkg/common/config/fuse.go
@@ -1,43 +1,0 @@
-/*
-Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserve.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package config
-
-import (
-	"os"
-)
-
-var defaultFuseConfig = FuseConfig{
-	EntryTimeout: 1,
-	AttrTimeout:  1,
-	Uid:          os.Getuid(),
-	Gid:          os.Getgid(),
-	RawOwner:     false,
-}
-
-type FuseConfig struct {
-	Uid          int
-	Gid          int
-	RawOwner     bool
-	EntryTimeout int
-	AttrTimeout  int
-}
-
-var FuseConf *FuseConfig
-
-func InitFuseConfig() {
-	FuseConf = &defaultFuseConfig
-}
