+++ pkg/fs/client/fuse/pfs.go
@@ -17,17 +17,33 @@ limitations under the License.
 package fuse
 
 import (
+	"os"
 	"time"
 
 	"github.com/hanwen/go-fuse/v2/fuse"
 	log "github.com/sirupsen/logrus"
 
-	"paddleflow/pkg/common/config"
 	"paddleflow/pkg/fs/client/meta"
 	"paddleflow/pkg/fs/client/vfs"
 )
 
 const fsName = "PaddleFlowFS"
+
+type FuseConfig struct {
+	Uid          int
+	Gid          int
+	RawOwner     bool
+	EntryTimeout int
+	AttrTimeout  int
+}
+
+var FuseConf = &FuseConfig{
+	EntryTimeout: 1,
+	AttrTimeout:  1,
+	Uid:          os.Getuid(),
+	Gid:          os.Getgid(),
+	RawOwner:     false,
+}
 
 type PFS struct {
 	debug bool
@@ -75,7 +91,7 @@ func (fs *PFS) GetAttr(cancel <-chan struct{}, input *fuse.GetAttrIn, out *fuse.
 		return fuse.Status(code)
 	}
 	attrToStat(entry.Ino, entry.Attr, &out.Attr)
-	out.AttrValid = uint64(config.FuseConf.AttrTimeout)
+	out.AttrValid = uint64(FuseConf.AttrTimeout)
 	if vfs.IsSpecialNode(meta.Ino(input.NodeId)) {
 		out.AttrValid = 3600
 	}
@@ -402,19 +418,19 @@ func (fs *PFS) replyEntry(entry *meta.Entry, out *fuse.EntryOut) {
 	out.NodeId = uint64(entry.Ino)
 	// todo:: Generation这个配置是干啥的，得在看看
 	out.Generation = 1
-	out.SetAttrTimeout(time.Duration(config.FuseConf.AttrTimeout))
+	out.SetAttrTimeout(time.Duration(FuseConf.AttrTimeout))
 	if entry.Attr.Type == meta.TypeDirectory {
 		// todo:: 增加dirEntry配置，目录和目录项超时分开设置
-		out.SetEntryTimeout(time.Duration(config.FuseConf.EntryTimeout))
+		out.SetEntryTimeout(time.Duration(FuseConf.EntryTimeout))
 	} else {
-		out.SetEntryTimeout(time.Duration(config.FuseConf.EntryTimeout))
+		out.SetEntryTimeout(time.Duration(FuseConf.EntryTimeout))
 	}
 	if vfs.IsSpecialNode(entry.Ino) {
 		out.SetAttrTimeout(time.Hour)
 	}
 	attrToStat(entry.Ino, entry.Attr, &out.Attr)
-	if !config.FuseConf.RawOwner {
-		out.Uid = uint32(config.FuseConf.Uid)
-		out.Gid = uint32(config.FuseConf.Gid)
+	if !FuseConf.RawOwner {
+		out.Uid = uint32(FuseConf.Uid)
+		out.Gid = uint32(FuseConf.Gid)
 	}
 }
